pyproject.toml
[tool.poetry]
name = "tgintegration"
version = "1.0.0"
description = "An Integration Test Library for Telegram Messenger Bots on top of Pyrogram."
authors = ["Joscha Götzer <joscha.goetzer@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "==3.*,>=3.6"
typing-extensions = "^3.7.4"
pyrogram = "^1.0.7"

[tool.poetry.dev-dependencies]
bumpversion = "==0.5.3"
coverage = "*"
tgcrypto = "==1.*,>=1.2.0"
python-decouple = "^3.3"
pytest = "*"
pytest-runner = "*"
pytest-asyncio = "*"
mkdocs-material = { version = "^6.0.2", python = "^3.7" }
mkapi = { version = "^1.0.13", python = "^3.7" }
poethepoet = "^0.9.0"
mkdocstrings = "^0.13.6"
mkdocs-toc-sidebar-plugin = "^0.1.0"

[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"

[tool.poe.tasks]
# Usage: `pip install poe`, then `poe [task]`
test = "pytest"
pre-stage-config = "git stage .pre-commit-config.yaml"
pre-install = "pre-commit install"
pre-commit = "pre-commit run --all-files"
pre = ["pre-stage-config", "pre-commit"]
pre-with-install = ["pre-stage-config", "pre-install", "pre-commit"]
docs-build = "mkdocs build"
docs-serve = "mkdocs serve"
docs-open = { script = "scripts.open_browser:open_docs" }
docs = ["docs-serve"]
